adalars-package {adalars}	R Documentation

The adaptive LASSO for linear regression ~~ package title ~~

Description

Use the adaptive lasso to find a shrunk-and-selected linear model for data with some censoring of the response variable. ~~ A concise (1-5 lines) description of the package ~~

Details

Package: 	 adalars

Type: 	 Package

Version: 	 0.1

Date: 	 2012-08-01

License: 	 GPL-3

LazyLoad: 	 yes

~~ An overview of how to use the package, including the most important functions ~~

Author(s)

Wesley Brooks

Maintainer: Wesley Brooks<wrbrooks@usgs.gov> ~~ The author and/or maintainer of the package ~~

[Package adalars version 0.1 Index]