adalars_step {adalars}	R Documentation

Uses the lars package to select variables based on the imputed observations that come out of the censReg step.

Usage

adalars_step(y, x)

Arguments

y

x

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(y, x) {
    result = list()
    
    p.max = min(dim(x)[2], dim(x)[1]-5)
    result[['cv']] = cv.model = cv.lars(y=y, x=x, type='lar', index=1:p.max, K=dim(x)[1])
    err.min = min(cv.model$cv)
    err.tol = err.min + cv.model$cv.error[which.min(cv.model$cv)]
    
    which.tol = which(cv.model$cv<err.tol)
    
    result[['lambda.index']] = lambda.index = min(which.tol)
    result[['model']] = model = lars(y=y, x=x, type='lar', max.steps=p.max)
    result[['fitted']] = predict.lars(model, newx=x, type='fit', s=lambda.index, mode='step')
    result[['vars']] = vars = names(which(model$beta[lambda.index,]>0))
    coefs = predict.lars(model, type='coefficients', s=lambda.index, mode='step')
    result[['coefs']] = coefs$coefficients[which(coefs$coefficients>0)]
    
    return(result)
  }

[Package adalars version 0.1 Index]